backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-simpleui>=2025.6.24",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "openai>=2.7.2",
    "pillow>=12.0.0",
    "pydub>=0.25.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
]

[[tool.uv.index]]
name = "aliyun"
url  = "https://mirrors.aliyun.com/pypi/simple/"
